another commit

+++ admin/includes/view_all_posts.php
@@ -20,6 +20,25 @@ if (isset($_POST['checkBoxArray'])) {
                 $update_to_delete_status = mysqli_query($connection, $query);
                 confirmQuery($update_to_delete_status);
                 break;
+
+            case 'clone':
+                $query = "SELECT * FROM posts where post_id = '{$postValueId}'";
+                $select_posts = mysqli_query($connection, $query);
+                while ($row = mysqli_fetch_assoc($select_posts)) {
+                    $post_id = $row['post_id'];
+                    $post_author = $row['post_author'];
+                    $post_title = $row['post_title'];
+                    $post_category_id = $row['post_category_id'];
+                    $post_status = $row['post_status'];
+                    $post_image = $row['post_image'];
+                    $post_tags = $row['post_tags'];
+                    $post_content = $row['post_content'];
+                    $post_comment_count = $row['post_comment_count'];
+                    $post_date = $row['post_date'];
+                }
+                $query = "INSERT INTO posts values(null,$post_category_id,'$post_title','$post_author', now(),'$post_image','$post_content','$post_tags','$post_comment_count','$post_status')";
+                $create_post_query = mysqli_query($connection, $query);
+                break;
         }
     }
 }
@@ -36,6 +55,7 @@ if (isset($_POST['checkBoxArray'])) {
                 <option value="published">Publish</option>
                 <option value="draft">Draft</option>
                 <option value="delete">Delete</option>
+                <option value="clone">Clone</option>
             </select>
         </div>
         <div class="col-xs-4">
@@ -57,12 +77,13 @@ if (isset($_POST['checkBoxArray'])) {
                 <th>View Post</th>
                 <th>Edit</th>
                 <th>Delete</th>
+                <th>Views</th>
             </tr>
         </thead>
         <tbody>
 
             <?php
-            $query = "SELECT * FROM posts";
+            $query = "SELECT * FROM posts ORDER BY post_id DESC";
             $select_posts = mysqli_query($connection, $query);
             while ($row = mysqli_fetch_assoc($select_posts)) {
                 $post_id = $row['post_id'];
@@ -74,6 +95,7 @@ if (isset($_POST['checkBoxArray'])) {
                 $post_tags = $row['post_tags'];
                 $post_comment_count = $row['post_comment_count'];
                 $post_date = $row['post_date'];
+                $post_views_count = $row['post_views_count'];
 
 
                 echo "<tr>";
@@ -105,7 +127,8 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td>$post_date</td>";
                 echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a onclick=\"javascript: return confirm('Are you sure you want to delete?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td>$post_views_count</td>";
                 echo "</tr>";
             }
             ?>
